Re-split frontends into Android and Desktop with a shared module
