+++ samples/emoji-search/android-composeui/build.gradle
@@ -28,12 +28,12 @@ dependencies {
   implementation projects.samples.emojiSearch.launcher
   implementation projects.samples.emojiSearch.presenterTreehouse
   implementation projects.samples.emojiSearch.schema.widget.protocol
+  implementation projects.samples.emojiSearch.sharedComposeui
   implementation libs.androidx.activity.compose
   implementation libs.androidx.appCompat
   implementation libs.androidx.core
   implementation libs.jetbrains.compose.material
   implementation libs.jetbrains.compose.ui
-  implementation libs.jetbrains.compose.ui.tooling.preview
   debugImplementation libs.jetbrains.compose.ui.tooling
   implementation projects.redwoodLayoutComposeui
   implementation projects.redwoodLazylayoutComposeui
