+++ samples/emoji-search/android-composeui/src/main/res/drawable/ic_launcher_background.xml
