+++ samples/emoji-search/android-composeui/src/main/res/mipmap-hdpi/ic_launcher.webp
