+++ samples/emoji-search/android-composeui/src/main/res/mipmap-mdpi/ic_launcher.webp
