+++ samples/emoji-search/android-composeui/src/main/res/mipmap-xhdpi/ic_launcher.webp
