+++ samples/emoji-search/android-composeui/src/main/res/mipmap-xxhdpi/ic_launcher.webp
