+++ samples/emoji-search/android-composeui/src/main/res/values/strings.xml
