--- samples/emoji-search/composeui/src/androidMain/AndroidManifest.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest
-  xmlns:android="http://schemas.android.com/apk/res/android">
-
-  <uses-permission android:name="android.permission.INTERNET" />
-  <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-
-  <application
-    android:name=".EmojiSearchApplication"
-    android:allowBackup="false"
-    android:exported="true"
-    android:icon="@mipmap/ic_launcher"
-    android:label="@string/app_name"
-    android:supportsRtl="true"
-    android:theme="@style/Theme.MyApplication"
-    android:usesCleartextTraffic="true"
-    >
-    <activity
-      android:name=".EmojiSearchActivity"
-      android:exported="true"
-      android:theme="@style/Theme.MyApplication.NoActionBar"
-      >
-      <intent-filter>
-        <action android:name="android.intent.action.MAIN" />
-        <category android:name="android.intent.category.DEFAULT" />
-        <category android:name="android.intent.category.LAUNCHER" />
-      </intent-filter>
-    </activity>
-  </application>
-</manifest>
