+++ samples/emoji-search/desktop-composeui/build.gradle
@@ -1,0 +1,30 @@
+apply plugin: 'org.jetbrains.kotlin.jvm'
+apply plugin: 'org.jetbrains.compose'
+
+compose {
+  desktop {
+    application {
+      mainClass = "com.example.redwood.emojisearch.desktop.Main"
+    }
+  }
+}
+
+dependencies {
+  implementation compose.desktop.currentOs
+  implementation projects.samples.emojiSearch.launcher
+  implementation projects.samples.emojiSearch.presenter
+  implementation projects.samples.emojiSearch.presenterTreehouse
+  implementation projects.samples.emojiSearch.schema.widget.protocol
+  implementation projects.samples.emojiSearch.sharedComposeui
+  implementation projects.redwoodComposeui
+  implementation projects.redwoodLayoutComposeui
+  implementation projects.redwoodLazylayoutComposeui
+  implementation projects.redwoodTreehouseHost
+  implementation projects.redwoodTreehouseHostComposeui
+  implementation projects.redwoodWidgetCompose
+  implementation libs.coil.compose
+  implementation libs.coil.network
+  implementation libs.ktor.engine.okhttp
+  implementation libs.zipline.loader
+  implementation libs.kotlinx.coroutines.swing
+}
